ENH: Add icon URL to API results
Change midas.slicerpackages.extension.list API to also include the icon
URL in the result metadata. We want this so that Slicer can use icons
for installed extensions. (It's not entirely decided if Slicer will
download the icons itself, or if we'll start packaging the icons in the
built extension packages, but either way this provides a neat solution
to Slicer needing to know the name - or at least the file extension - of
the icon.)